move all fns into utils dir
