+++ packages/babel-plugin/src/utils/buildClassesNode.js
