+++ packages/babel-plugin/src/utils/extendOptions.js
