+++ packages/babel-plugin/src/utils/findStylesNode.js
