+++ packages/babel-plugin/src/utils/isSupportedCallIdentifier.js
