+++ packages/babel-plugin/src/utils/removeEmptyObjects.js
