+++ packages/babel-plugin/src/utils/removeNonFunctionProps.js
