+++ packages/babel-plugin/src/utils/resolveRef.js
