+++ packages/babel-plugin/src/utils/serializeNode.js
