+++ packages/babel-plugin/src/utils/serializePropAccess.js
